CyberPet/requirements.txt
flask==2.3.3
inky==1.4.0
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.8.4
reportlab==4.0.9
tensorflow==2.15.0
pybluez==0.23